Fix arithmetic overflow when calculating Fibonacci numbers with index bigger than 92

+++ tests/Fibonacci.Tests/Solver/TestFibonacciCache.cs
@@ -1,18 +1,19 @@
 ﻿using System.Collections.Generic;
+using System.Numerics;
 using Fibonacci.Cache;
 
 namespace Fibonacci.Tests.Solver
 {
     public class TestFibonacciCache : IFibonacciCache
     {
-        private readonly Dictionary<int, ulong> _cache = new Dictionary<int, ulong>();
+        private readonly Dictionary<int, BigInteger> _cache = new Dictionary<int, BigInteger>();
 
-        public void Save(int index, ulong value)
+        public void Save(int index, BigInteger value)
         {
             _cache.Add(index, value);
         }
 
-        public (bool Exists, ulong Value) Retrieve(int index)
+        public (bool Exists, BigInteger Value) Retrieve(int index)
         {
             var exists = _cache.TryGetValue(index, out var value);
             return (exists, value);
